pyproject.toml
[project]
name = "jotunn"
version = "0.1.261"
description = "A petabyte scale data processing framework for AI models using Ray."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Teraflop AI", email = "enrico@teraflop.ai"},
]
classifiers = [
  "Typing :: Typed",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "daft",
    "google-re2>=1.1.20250805",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "memray>=1.18.0",
    "open-clip-torch>=3.2.0",
    "pysimdjson>=7.0.2",
    "rapidfuzz>=3.14.1",
    "ray[data]>=2.47.1",
    "transformers>=4.53.3",
    "usearch>=2.21.0",
    "xxhash>=3.5.0",
]
[project.optional-dependencies]
text = [
    "trafilatura>=2.0.0",
    "chonkie>=1.1.0",
    "sentence-transformers>=5.0.0",
    "vllm>=0.9.2",
]
audio = [
    "librosa>=0.11.0",
    "silero-vad>=5.1.2",
    "soundfile>=0.13.1",
]
video = [
    "av>=15.1.0",
    "scenedetect>=0.6.6",
    "video-reader-rs>=0.2.10",
]
image = [
    "easyocr>=1.7.2",
    "exifread>=3.5.1",
    "imagehash>=4.3.2",
    "opencv-python>=4.12.0.88",
    "pillow-simd>=9.5.0.post2",
    "scikit-image>=0.25.2",
    "simple-aesthetics-predictor>=0.1.2",
    "vllm>=0.9.2",
]
anyscale = [
    "anyscale>=0.26.58",
]
all = [
    "jotunn[text,audio,video,image,anyscale]",
]

[tool.uv]
package = true
